Adding IniialMigrations and SetupRequiredandUnique
Delete 2 method ToViewModel and ToModel inside UserViewModel.cs
Adding Name is Unique in file ApplicationDbContext
Change OrderOfImportance is not required
Add Required name for Group
Change ActionResult<T> to IActionResult

+++ SecretSanta/src/SecretSanta.Api/Controllers/UserController.cs
@@ -5,40 +5,57 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using SecretSanta.Api.ViewModels;
 using SecretSanta.Domain.Services.Interfaces;
+using Microsoft.AspNetCore.Http;
+using AutoMapper;
+using SecretSanta.Domain.Models;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace SecretSanta.Api.Controllers
 {
     [Route("api/[controller]")]
+    [ApiController]
     public class UserController : ControllerBase
     {
         private IUserService UserService { get; }
 
-        public UserController(IUserService userService)
+        private IMapper Mapper { get; }
+  
+        public UserController(IUserService userService, IMapper mapper)
         {
-            UserService = userService;
+            UserService = userService ?? throw new ArgumentNullException(nameof(userService));
+            Mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
-        // POST api/<controller>
+        // GET api/User
+        [HttpGet]
+        [Produces(typeof(ICollection<UserViewModel>))]
+        public IActionResult Get()
+        {
+            return Ok(UserService.FetchAll().Select(x => Mapper.Map<UserViewModel>(x)));
+        }
+
+        // POST api/User
         [HttpPost]
-        public ActionResult<UserViewModel> Post(UserInputViewModel userViewModel)
+        [Produces(typeof(UserViewModel))]
+        public IActionResult Post(UserInputViewModel userViewModel)
         {
             if (userViewModel == null)
             {
                 return BadRequest();
             }
-
-            var persistedUser = UserService.AddUser(UserInputViewModel.ToModel(userViewModel));
-
-            return Ok(UserViewModel.ToViewModel(persistedUser));
+       
+            var createdUser = UserService.AddUser(Mapper.Map<User>(userViewModel));
+       
+            return CreatedAtAction(nameof(Get), new { id = createdUser.Id }, createdUser);
+            
         }
 
         // PUT api/<controller>/5
-        [HttpPut("{id}")]
-        public ActionResult<UserViewModel> Put(int id, UserInputViewModel userViewModel)
+        [HttpPut("{id}")]        
+        public IActionResult Put(int id, UserInputViewModel viewModel)
         {
-            if (userViewModel == null)
+            if (viewModel == null)
             {
                 return BadRequest();
             }
@@ -49,21 +66,38 @@ namespace SecretSanta.Api.Controllers
                 return NotFound();
             }
 
-            foundUser.FirstName = userViewModel.FirstName;
-            foundUser.LastName = userViewModel.LastName;
-
-            var persistedUser = UserService.UpdateUser(foundUser);
-
-            return Ok(UserViewModel.ToViewModel(persistedUser));
+            Mapper.Map(viewModel, foundUser);
+            UserService.UpdateUser(foundUser);
+            
+            return NoContent();
         }
 
         // DELETE api/<controller>/5
         [HttpDelete("{id}")]
-        public ActionResult Delete(int id)
+        public IActionResult Delete(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("A User id must be specified");
+            }
+
             bool userWasDeleted = UserService.DeleteUser(id);
 
             return userWasDeleted ? (ActionResult)Ok() : (ActionResult)NotFound();
+        }
+
+        // GET api/<controller>/5
+        [HttpGet("{id}")]
+        [Produces(typeof(UserViewModel))]
+        public IActionResult Get(int id)
+        {
+            var foundUser = UserService.Find(id);
+            if (foundUser == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(Mapper.Map<UserViewModel>(foundUser));
         }
     }
 }
